+++ fuzzing_cli/fuzz/ide/brownie.py
@@ -2,7 +2,7 @@ import logging
 import os
 from functools import lru_cache
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Optional, Set, Tuple
 
 from fuzzing_cli.fuzz.config import FuzzingOptions
 from fuzzing_cli.fuzz.exceptions import BuildArtifactsError
@@ -121,3 +121,7 @@ class BrownieArtifacts(IDEArtifacts):
                         f"Build artifact did not contain expected key. Contract: {contract}: \n{e}"
                     )
         return result_contracts, result_sources
+
+    def unlinked_libraries(self) -> List[Tuple[Contract, Set[str]]]:
+        # We'll deprecate Brownie in the future, so we don't implement this method
+        return []